semi working code using OleMenuCommand
The first version uses the community vsix helper library but that did
not work with non-code type files, e.g. PNGs.  This uses a similar
technique found in the FindInSolutionExplorer package.

Specifically, it uses the trick of turning on file selection tracking in
solution explorer temporarily so VS does the work and then giving
focus to Solution Explorer.  However, in VS2022, there seems to be
some kind of race condition as it only works abou 80% of the time.