can execute check relies on project name instead
Prior, the OnMenuCommandBeforeQueryStatus method used only the caption
of the Active window to determine if it command can execute.  But this
would fail on windows that contained sub-windows, like a .net 4.x csproj
file because the focus would usually not be on the main tab window.

This change implements a two step check.  First it will check for the
Project property of the ActiveWindow.  If it is present, it will use its
name to determine if the command can be enabled.  If it is null,
as is the case with tabs with sub-windows that are in the solutions, it
will fall back to the ActiveWindow caption.  If that is null or not
located in the tree, it will disable the control.