fix context menu icon disable on selectable file
If the user goes from a non selectable tab to right-clicking on one that
is, does not re-enable.  This is due to the method that sets Enabled on
the command being performed on a dispatched Task.  This must be
done synchronously so the setting is ready before the context menu
is rendered.